Move molecules scenarios
