+++ inductiva/molecules/__init__.py
@@ -1,3 +1,4 @@
 #pylint: disable=missing-module-docstring
-from .scenarios import ProteinSolvation
-from .scenarios import MDWaterBox
+from .protein_solvation import ProteinSolvation, ProteinSolvationOutput
+from .md_water_box import MDWaterBox, MDWaterBoxOutput
+from . import utils
