--- inductiva/molecules/scenarios/__init__.py
@@ -1,4 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .protein_solvation import ProteinSolvation, ProteinSolvationOutput
-from .md_water_box import MDWaterBox, MDWaterBoxOutput
-from . import utils
